+++ concour_main.pas
@@ -231,7 +231,7 @@ begin
   Result := -1;
   RouteTypeSL.BeginUpdate;
   for i:=0 to RouteTypeSL.Count-1 do
-    if Integer(RouteTypeSL.Objects[i])=RType then
+    if Int64(RouteTypeSL.Objects[i])=RType then
     begin
       Result := i;
       Break;
@@ -887,7 +887,7 @@ begin
     {было    if (DM.CurrRouteType>=0) and (DM.CurrRouteType<RouteTypeSL.Count-1) then
                        StatusBar1.Panels[1].Text:= RouteTypeSL[DM.CurrRouteType];
     }
-    x := RouteTypeSL.IndexOfObject(TObject(DM.CurrRouteType));
+    x := RouteTypeSL.IndexOfObject(TObject(Int64(DM.CurrRouteType)));
     if x>0 then StatusBar1.Panels[1].Text:= RouteTypeSL[x];
     //-
     // в диалоге выбора уже открыли: DM.OpenRoutes(DM.CurrentRoute);
